Changes, Minor optimizations an Language typo fixes

+++ Blitz Troubleshooter/Blitz Troubleshooter V2.3 + V2.4/Blitz Troubleshooter V2.3 Solution/Blitz Troubleshooter V2.3/English.xaml.cs
@@ -20,55 +20,51 @@ namespace Blitz_Troubleshooter_V2._3
 
         public void DownloadVSRDST()
         {
-
-
             using (var client = new WebClient())
             {
 
-                client.DownloadFile("https://aka.ms/vs/16/release/vc_redist.x86.exe", System.IO.Path.GetTempPath() + "vc_redist.x86.exe");
+                try
+                {
+                    client.DownloadFile("https://aka.ms/vs/16/release/vc_redist.x86.exe", System.IO.Path.GetTempPath() + "vc_redist.x86.exe");
+                }
+                catch (Exception e)
+                {
+                    MessageBox.Show(e.Message);
+                    System.Windows.Application.Current.Shutdown();
+                }
+                finally
+                {
+                    Process.Start(System.IO.Path.GetTempPath() + "vc_redist.x86.exe");
+                }
             }
-
-
 
         }
         public void KillBlitz()
         {
-            try
+            foreach (Process proc in Process.GetProcessesByName("Blitz"))
             {
-                foreach (Process proc in Process.GetProcessesByName("Blitz"))
-                {
-                    proc.Kill();
-                }
+                proc.Kill();
             }
-            catch (Exception ex)
-            {
-                MessageBox.Show(ex.Message);
-            }
+
         }
         public void AppdataBlitz()
         {
-            try
+            var path = Environment.GetFolderPath(Environment.SpecialFolder.ApplicationData) + "\\Blitz";
+            if (Directory.Exists(path))
             {
-                var path = Environment.GetFolderPath(Environment.SpecialFolder.ApplicationData) + "\\Blitz";
                 Directory.Delete(path, true);
-
             }
-            catch (Exception e)
-            {
-                MessageBox.Show(e.Message);
-            }
-
         }
         public void Uninstall()
         {
             var paths = new String[] {
-            Environment.GetFolderPath(Environment.SpecialFolder.LocalApplicationData) + "\\blitz-updater",
-            Environment.GetFolderPath(Environment.SpecialFolder.LocalApplicationData) + "\\Programs\\blitz-core",
-            Environment.GetFolderPath(Environment.SpecialFolder.ApplicationData) + "\\Blitz",
-            Environment.GetFolderPath(Environment.SpecialFolder.ApplicationData) + "\\blitz-core",
-            Environment.GetFolderPath(Environment.SpecialFolder.ApplicationData) + "\\Blitz-helpers",
-            Environment.GetFolderPath(Environment.SpecialFolder.LocalApplicationData) + "\\Programs\\Blitz"
-           };
+        Environment.GetFolderPath(Environment.SpecialFolder.LocalApplicationData) + "\\blitz-updater",
+        Environment.GetFolderPath(Environment.SpecialFolder.LocalApplicationData) + "\\Programs\\blitz-core",
+        Environment.GetFolderPath(Environment.SpecialFolder.ApplicationData) + "\\Blitz",
+        Environment.GetFolderPath(Environment.SpecialFolder.ApplicationData) + "\\blitz-core",
+        Environment.GetFolderPath(Environment.SpecialFolder.ApplicationData) + "\\Blitz-helpers",
+        Environment.GetFolderPath(Environment.SpecialFolder.LocalApplicationData) + "\\Programs\\Blitz"
+      };
 
             foreach (var path in paths)
             {
@@ -84,94 +80,69 @@ namespace Blitz_Troubleshooter_V2._3
         public void CleanReinstall()
         {
             var paths = new String[] {
-            Environment.GetFolderPath(Environment.SpecialFolder.LocalApplicationData) + "\\blitz-updater",
-            Environment.GetFolderPath(Environment.SpecialFolder.LocalApplicationData) + "\\Programs\\blitz-core",
-            Environment.GetFolderPath(Environment.SpecialFolder.ApplicationData) + "\\Blitz",
-            Environment.GetFolderPath(Environment.SpecialFolder.ApplicationData) + "\\blitz-core",
-            Environment.GetFolderPath(Environment.SpecialFolder.ApplicationData) + "\\Blitz-helpers",
-            Environment.GetFolderPath(Environment.SpecialFolder.LocalApplicationData) + "\\Programs\\Blitz"
-           };
+        Environment.GetFolderPath(Environment.SpecialFolder.LocalApplicationData) + "\\blitz-updater",
+        Environment.GetFolderPath(Environment.SpecialFolder.LocalApplicationData) + "\\Programs\\blitz-core",
+        Environment.GetFolderPath(Environment.SpecialFolder.ApplicationData) + "\\Blitz",
+        Environment.GetFolderPath(Environment.SpecialFolder.ApplicationData) + "\\blitz-core",
+        Environment.GetFolderPath(Environment.SpecialFolder.ApplicationData) + "\\Blitz-helpers",
+        Environment.GetFolderPath(Environment.SpecialFolder.LocalApplicationData) + "\\Programs\\Blitz"
+      };
 
-            try
+            foreach (var path in paths)
             {
-                foreach (var path in paths)
+                if (Directory.Exists(path))
                 {
-                    if (Directory.Exists(path))
-                    {
-                        Directory.Delete(path, true);
-                    }
+                    Directory.Delete(path, true);
                 }
             }
-            catch { }
-            finally
+
+            System.Threading.Thread.Sleep(5000);
+
+            MessageBox.Show("Please wait while Blitz completely reinstalls itself", "Blitz Clean Reinstallation", MessageBoxButton.OK, MessageBoxImage.Information);
+            using (var client = new WebClient())
             {
-
-
-                System.Threading.Thread.Sleep(5000);
-
-
-
-                MessageBox.Show("Please wait while Blitz completely reinstalls itself", "Blitz Clean Reinstallation", MessageBoxButton.OK, MessageBoxImage.Information);
-                using (var client = new WebClient())
+                try
                 {
-                    try
-                    {
-                        client.DownloadFile("https://blitz.gg/download/win", System.IO.Path.GetTempPath() + "Blitz.exe");
-                    }
-                    catch (Exception e)
-                    {
-                        MessageBox.Show(e.Message);
-                        System.Windows.Application.Current.Shutdown();
-                    }
-                    finally
-                    {
-                        Process.Start(System.IO.Path.GetTempPath() + "Blitz.exe");
-                    }
+                    client.DownloadFile("https://blitz.gg/download/win", System.IO.Path.GetTempPath() + "Blitz.exe");
                 }
-
-
-
+                catch (Exception e)
+                {
+                    MessageBox.Show(e.Message);
+                    System.Windows.Application.Current.Shutdown();
+                }
+                finally
+                {
+                    Process.Start(System.IO.Path.GetTempPath() + "Blitz.exe");
+                }
             }
-
         }
 
         private void Button_Click(object sender, RoutedEventArgs e)
         {
-            //fix all commmon mistakes
             KillBlitz();
             System.Threading.Thread.Sleep(1000);
             CleanReinstall();
-
-
         }
 
         private void Button_Click_1(object sender, RoutedEventArgs e)
         {
             KillBlitz();
             DownloadVSRDST();
-            Process.Start(System.IO.Path.GetTempPath() + "vc_redist.x86.exe");
             MessageBox.Show("Please proceed to the other Window, and finish the installation, press the OK button to continue.", "vc_redist.x86 installation", MessageBoxButton.OK, MessageBoxImage.Information);
-
-        } /*Done */
+        }
 
         private void Button_Click_2(object sender, RoutedEventArgs e)
         {
-            //Fix Cache issues
             KillBlitz();
             System.Threading.Thread.Sleep(1000);
             AppdataBlitz();
-            MessageBox.Show("Cache sucessfully cleared");
-        } /*Done */
-
-
-
+            MessageBox.Show("Cache successfully cleared");
+        }
         private void Button_Click_4(object sender, RoutedEventArgs e)
         {
             KillBlitz();
             System.Threading.Thread.Sleep(1000);
             Uninstall();
         }
-
-
     }
 }
