+++ Blitz Troubleshooter/Blitz Troubleshooter V2.3 + V2.4/Blitz Troubleshooter V2.3 Solution/Blitz Troubleshooter V2.3/German.xaml.cs
@@ -14,46 +14,46 @@ namespace Blitz_Troubleshooter_V2._3
         public German(string version)
         {
             InitializeComponent();
+
             w1.Title = "Blitz Troubleshooter V" + version;
         }
 
         public void DownloadVSRDST()
         {
-
             using (var client = new WebClient())
             {
 
-                client.DownloadFile("https://aka.ms/vs/16/release/vc_redist.x86.exe", System.IO.Path.GetTempPath() + "vc_redist.x86.exe");
+                try
+                {
+                    client.DownloadFile("https://aka.ms/vs/16/release/vc_redist.x86.exe", System.IO.Path.GetTempPath() + "vc_redist.x86.exe");
+                }
+                catch (Exception e)
+                {
+                    MessageBox.Show(e.Message);
+                    System.Windows.Application.Current.Shutdown();
+                }
+                finally
+                {
+                    Process.Start(System.IO.Path.GetTempPath() + "vc_redist.x86.exe");
+                }
             }
 
         }
         public void KillBlitz()
         {
-            try
+            foreach (Process proc in Process.GetProcessesByName("Blitz"))
             {
-                foreach (Process proc in Process.GetProcessesByName("Blitz"))
-                {
-                    proc.Kill();
-                }
+                proc.Kill();
             }
-            catch (Exception ex)
-            {
-                MessageBox.Show(ex.Message);
-            }
+
         }
         public void AppdataBlitz()
         {
-            try
+            var path = Environment.GetFolderPath(Environment.SpecialFolder.ApplicationData) + "\\Blitz";
+            if (Directory.Exists(path))
             {
-                var path = Environment.GetFolderPath(Environment.SpecialFolder.ApplicationData) + "\\Blitz";
                 Directory.Delete(path, true);
-
             }
-            catch (Exception e)
-            {
-                MessageBox.Show(e.Message);
-            }
-
         }
         public void Uninstall()
         {
@@ -66,19 +66,15 @@ namespace Blitz_Troubleshooter_V2._3
         Environment.GetFolderPath(Environment.SpecialFolder.LocalApplicationData) + "\\Programs\\Blitz"
       };
 
-            try
+            foreach (var path in paths)
             {
-                foreach (var path in paths)
+                if (Directory.Exists(path))
                 {
-                    if (Directory.Exists(path))
-                    {
-                        Directory.Delete(path, true);
-                    }
+                    Directory.Delete(path, true);
                 }
             }
-            catch { }
 
-            MessageBox.Show("Blitz wurde deinstalliert ", " Blitz-Deinstallation ", MessageBoxButton.OK, MessageBoxImage.Information);
+            MessageBox.Show("Blitz wurde deinstalliert ", " Blitz Deinstallation", MessageBoxButton.OK, MessageBoxImage.Information);
         }
 
         public void CleanReinstall()
@@ -92,42 +88,33 @@ namespace Blitz_Troubleshooter_V2._3
         Environment.GetFolderPath(Environment.SpecialFolder.LocalApplicationData) + "\\Programs\\Blitz"
       };
 
-            try
+            foreach (var path in paths)
             {
-                foreach (var path in paths)
+                if (Directory.Exists(path))
                 {
-                    if (Directory.Exists(path))
-                    {
-                        Directory.Delete(path, true);
-                    }
+                    Directory.Delete(path, true);
                 }
             }
-            catch { }
-            finally
+
+            System.Threading.Thread.Sleep(5000);
+
+            MessageBox.Show("Bitte warten Sie, während sich Blitz vollständig neu installiert ", " Blitz Clean Neuinstallation", MessageBoxButton.OK, MessageBoxImage.Information);
+            using (var client = new WebClient())
             {
-
-                System.Threading.Thread.Sleep(5000);
-
-                MessageBox.Show("Bitte warten Sie, während sich Blitz vollständig neu installiert ", " Blitz Clean Neuinstallation", MessageBoxButton.OK, MessageBoxImage.Information);
-                using (var client = new WebClient())
+                try
                 {
-                    try
-                    {
-                        client.DownloadFile("https://blitz.gg/download/win", System.IO.Path.GetTempPath() + "Blitz.exe");
-                    }
-                    catch (Exception e)
-                    {
-                        MessageBox.Show(e.Message);
-                        System.Windows.Application.Current.Shutdown();
-                    }
-                    finally
-                    {
-                        Process.Start(System.IO.Path.GetTempPath() + "Blitz.exe");
-                    }
+                    client.DownloadFile("https://blitz.gg/download/win", System.IO.Path.GetTempPath() + "Blitz.exe");
                 }
-
+                catch (Exception e)
+                {
+                    MessageBox.Show(e.Message);
+                    System.Windows.Application.Current.Shutdown();
+                }
+                finally
+                {
+                    Process.Start(System.IO.Path.GetTempPath() + "Blitz.exe");
+                }
             }
-
         }
 
         private void Button_Click(object sender, RoutedEventArgs e)
@@ -135,33 +122,27 @@ namespace Blitz_Troubleshooter_V2._3
             KillBlitz();
             System.Threading.Thread.Sleep(1000);
             CleanReinstall();
-
         }
 
         private void Button_Click_1(object sender, RoutedEventArgs e)
         {
             KillBlitz();
             DownloadVSRDST();
-            Process.Start(System.IO.Path.GetTempPath() + "vc_redist.x86.exe");
             MessageBox.Show("Fahren Sie mit dem anderen Fenster fort und beenden Sie die Installation. Klicken Sie auf OK, um fortzufahren. ", " Vc_redist.x86 Installation", MessageBoxButton.OK, MessageBoxImage.Information);
-
         }
 
         private void Button_Click_2(object sender, RoutedEventArgs e)
         {
-
             KillBlitz();
             System.Threading.Thread.Sleep(1000);
             AppdataBlitz();
             MessageBox.Show("Cache erfolgreich geleert");
         }
-
         private void Button_Click_4(object sender, RoutedEventArgs e)
         {
             KillBlitz();
             System.Threading.Thread.Sleep(1000);
             Uninstall();
         }
-
     }
 }
